% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_taxa_label}
\alias{make_taxa_label}
\title{Generate a unique taxonomic label from taxonomic ranks}
\usage{
make_taxa_label(df)
}
\arguments{
\item{df}{A dataframe containing at least the columns: Kingdom, Phylum, Class, Order, Family, Genus.
If \code{Species} is present, it will be used as the terminal label.}
}
\value{
The input dataframe with an added \code{Taxa} column
}
\description{
Creates a \code{Taxa} column in the dataframe based on taxonomic hierarchy.
If \code{Species} column is present and not missing, it is used as the most specific label.
Otherwise, the function uses \code{Genus}, or the highest classified rank available.
}
