% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutt-package.R
\docType{package}
\name{mutt-package}
\alias{mutt-package}
\title{MUTT package internals}
\description{
Parses and formats individual studies containing microbiome sequence count data paired with total abundance measurements (flow cytometry, ddPCR, qPCR, etc.), and returns a single data object containing selected studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Silverman-Lab/mutt}
  \item Report bugs at \url{https://github.com/Silverman-Lab/mutt/issues}
}

}
\author{
\strong{Maintainer}: Maxwell Konnaris \email{mak6930@psu.edu}

Authors:
\itemize{
  \item Maxwell Konnaris \email{mak6930@psu.edu}
  \item Justin Silverman \email{JustinSilverman@psu.edu} [contributor]
}

}
\keyword{internal}
