% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.mutt_result}
\alias{print.mutt_result}
\title{Print a MUTT result summary}
\usage{
\method{print}{mutt_result}(x, ...)
}
\arguments{
\item{x}{A result returned by \code{\link[=mutt]{mutt()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a MUTT result summary
}
