% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_empty_columns}
\alias{remove_empty_columns}
\title{Remove columns that are entirely NA, empty strings, or whitespace}
\usage{
remove_empty_columns(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with empty columns removed
}
\description{
This function removes columns from a data frame where all values are either:
\itemize{
\item NA
\item Empty strings ("")
\item Whitespace only (" ")
}
}
\examples{
\dontrun{
# Remove empty columns from a data frame
df <- data.frame(
  A = c(1, 2, 3),
  B = c("", "", ""),
  C = c(NA, NA, NA),
  D = c(" ", " ", " ")
)
clean_df <- remove_empty_columns(df)
}

}
