% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{standardize_output_order}
\alias{standardize_output_order}
\title{Standardize the order of elements in a dataset output}
\usage{
standardize_output_order(output)
}
\arguments{
\item{output}{A list containing the dataset output}
}
\value{
The same list with elements reordered to the standard order
}
\description{
Reorders the elements of a dataset output to follow the standard order:
counts, proportions, tax, scale, metadata, followed by any additional elements
}
\examples{
\dontrun{
# Standardize the order of elements in a dataset output
standardized_output <- standardize_output_order(my_output)
}

}
